102-concat.js

+++ 0x13-javascript_objects_scopes_closures/102-concat.js
@@ -3,11 +3,13 @@ const av = process.argv.slice(2);
 const ac = av.length;
 if (ac >= 3) {
   const fs = require('fs');
-  try {
+  if (fs.existsSync(av[0]) &&
+      fs.statSync(av[0]).isFile &&
+      fs.existsSync(av[1]) &&
+      fs.statSync(av[1]).isFile) {
     const contentA = fs.readFileSync(av[0], 'utf8').trim();
     const contentB = fs.readFileSync(av[1], 'utf8').trim();
     fs.writeFileSync(av[2], contentA + '\n');
     fs.appendFileSync(av[2], contentB + '\n');
-  } catch (err) {
   }
 }
